EC_SPACE/requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==4.2
django-restframework==0.0.1
idna==3.10
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
